\docType{class}
\name{Contacts-class}
\alias{Contacts-class}
\title{Class \code{"Contacts"}}
\description{
  Class to handle contacts.

  Class to handle contacts.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("Contacts", root, startDate, days, source,
  destination, t, id, n, category, level, direction, ...)}.
}

\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("Contacts", root, startDate, days, source,
  destination, t, id, n, category, level, direction, ...)}.
}
\examples{
showClass("Contacts")
showClass("Contacts")
}
\keyword{classes}

