\docType{class}
\name{ContactTrace-class}
\alias{ContactTrace-class}
\title{Class \code{"ContactTrace"}}
\description{
  Class to handle contact tracing.

  Class to handle contact tracing.
}
\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("ContactTrace", root, ingoingContacts,
  outgoingContacts, ...)}. %% ~~ describe objects here ~~
}

\section{Objects from the Class}{
  Objects can be created by calls of the form
  \code{new("ContactTrace", root, ingoingContacts,
  outgoingContacts, ...)}. %% ~~ describe objects here ~~
}
\examples{
showClass("ContactTrace")
showClass("ContactTrace")
}
\keyword{classes}

