\name{position_tree}
\alias{position_tree}
\title{Position nodes in a tree}
\usage{
  position_tree(tree, sibling_separation = 4,
    level_separation = 1, x_top_adjustment = 0,
    y_top_adjustment = 0)
}
\arguments{
  \item{tree}{}

  \item{sibling_separation}{}

  \item{x_top_adjustment}{}

  \item{y_top_adjustemnt}{}
}
\description{
  This function determines the coordinates for each node in
  a tree. A pointer to the apex node of the tree is passed
  as
}
\references{
  \itemize{ \item John Q. Walker II, A node positioning
  algorithm for general tress.\cr
  \url{http://www.cs.unc.edu/techreports/89-034.pdf} }
}
\keyword{internal}

