% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contacts.R
\docType{class}
\name{Contacts-class}
\alias{Contacts-class}
\title{Class \code{"Contacts"}}
\description{
Class to handle contacts.
}
\details{
The \code{Contacts} class keeps track of all ingoing or outgoing
livstock transfers in the contact chain for a specific root within
the time window used for contact tracing.  The slots;
\code{source}, \code{destination}, \code{t}, \code{id}, \code{n}
and \code{category} contains contact information extracted from
the movement dataset during contact tracing. The \code{index} slot
is an index to the extracted contacts within the class that
together with the \code{distance} slot can be used to rebuild the
exact contacts that were extracted from each search step during
the contact tracing.
}
\section{Slots}{

\describe{
\item{\code{root}}{A \code{character} vector of length one with the
identifier of the root.}

\item{\code{tBegin}}{A \code{Date} vector of length one with the start
date of the time window used for contact tracing.}

\item{\code{tEnd}}{A \code{Date} vector of length one with the end date of
the time window used for contact tracing.}

\item{\code{source}}{A \code{character} vector with the identifiers of the
source holdings of the livestock transfer.}

\item{\code{destination}}{A \code{character} vector with the identifier of
the destination holdings of the livestock transfer.}

\item{\code{t}}{A \code{Date} vector of the livestock transfer.}

\item{\code{id}}{A \code{character} vector with the identifiers of the
animals.}

\item{\code{n}}{A \code{numeric} vector with the number of animals
transfered.}

\item{\code{category}}{A \code{character} vector with the category of
animals e.g. cattle.}

\item{\code{index}}{A \code{integer} index vector.}

\item{\code{distance}}{A \code{integer} vector with the distance from root
for the contact[index]}

\item{\code{direction}}{A \code{character} vector of length one equal to
the direction "in" or "out"}
}}

\section{Objects from the Class}{
 Objects can be created by calls
    of the form \code{new("Contacts", root, startDate, days,
    source, destination, t, id, n, category, level, direction,
    ...)}.
}

\examples{

## Load data
data(transfers)

## Perform contact tracing
contactTrace <- Trace(movements = transfers,
                      root = 2645,
                      tEnd = "2005-10-31",
                      days = 90)

## Show structure of ingoing contacts
str(contactTrace@ingoingContacts)

## Show structure of ougoing contacts
str(contactTrace@outgoingContacts)

}
