% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{ReportObject}
\alias{ReportObject}
\title{Get current \code{ContactTrace} object when generating a report}
\usage{
ReportObject()
}
\value{
The current \code{ContactTrace} object when generating a report
}
\description{
EpiContatTrace contains report templates to generate pdf- or html reports
for the farm specific contacts. These reports can be useful for hands-on
disease tracing in the field. The templates are used by Sweave and can be
adapted by the end user. This method enables communication of the current
\code{ContactTrace} object to the report.
}
